Add serial version

+++ source/module_pw/pw_transform.cpp
@@ -206,7 +206,7 @@ void PW_Basis:: recip2real(std::complex<float> * in, std::complex<float> * out)
     
     for(int ir = 0 ; ir < this->nrxx ; ++ir)
     {
-        out[ir] = this->ft.auxf1[ir] / double(this->bignxyz);
+        out[ir] = this->ft.auxf1[ir] / float(this->bignxyz);
     }
 
     return;
@@ -232,7 +232,7 @@ void PW_Basis:: recip2real(std::complex<float> * in, float * out)
     {
         for(int ipy = 0 ; ipy < npy ; ++ipy)
         {
-            out[ix*npy + ipy] = this->ft.rf_rspace[ix*npy*2 + ipy] / double(this->bignxyz);
+            out[ix*npy + ipy] = this->ft.rf_rspace[ix*npy*2 + ipy] / float(this->bignxyz);
         }
     }
     return;
